Backend/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymssql>=2.3.1
bcrypt==4.1.2
pyjwt==2.8.0
